Rename variables and modules
